made the tests return 200 OK
